gcp/functions/pubsub_to_langsmith/requirements.txt
functions-framework==3.*
langsmith>=0.1.31
google-cloud-logging>=3.9.0
python-dotenv>=1.0.1
